# Faster hexdigest