--- a/Gemfile
+++ b/Gemfile
@@ -1,9 +1,10 @@
 source "https://rubygems.org"
-gem "rack", github: 'rack/rack'
 gemspec
 
 if Gem::Version.new(RUBY_VERSION) < Gem::Version.new("2.2.2")
   gem 'rack', '< 2.0'
+else
+  gem "rack", github: 'rack/rack'
 end
 
 if Gem::Version.new(RUBY_VERSION) < Gem::Version.new("2.0")