@schneems 2016-07-04
Did some more benchmarking and the returns diminish the more complicated the item being digested is. This shows a great improvement because "foo" is so simple. It's still faster when using a more complicated hash object, but not quite such a jump.